fix: separate file of arbitrary-annotation
