+++ pydtk/db/schemas/dataware-tools.com/v1alpha5/annotation.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-from pydantic import Extra, Field, constr
+from pydantic import Field, constr
 
 from pydtk.db.schemas import BaseSchema, register_schema
 
@@ -18,16 +18,3 @@ class Annotation(BaseSchema):
     timestamp_to: Union[float, None]
     created_at: Union[float, None]
     created_by: Union[str, None]
-
-
-@register_schema
-class ArbitraryAnnotation(Annotation):
-    """Schema for an annotation with extra fields."""
-
-    _api_version = "dataware-tools.com/v1alpha5"
-    _kind = "ArbitraryAnnotation"
-
-    class Config:
-        """Configs for ArbitraryAnnotation."""
-
-        extra = Extra.allow
